Refactor Exporters to Exportables
Instead of having dedicated exporters that loop through rules and construct FSH in the export, wrap SUSHI's FSH classes as "Exportable" subclasses that support a `toFHIR()` function.

To keep TypeScript happy, this required stubbing out a lot of the classes we haven't implemented yet -- but at least now we have a starting point!