Extract ContainsRule
A ContainsRule specifies information about an element that is a slice of
another element. Only elements that are slices should be given to the
ContainsRuleExtractor. Because a ContainsRule includes cardinality and
flag information, the element will not be processed by the
CardRuleExtractor or the FlagRuleExtractor.

A ContainsRule may have a type specified in addition to a name. However,
this is not necessary and is not currently implemented. The reason for
this is that a type can only be specified as part of the ContainsRule
when the slice has exactly one type with at most one profile. A TODO has
been added to Package.optimize regarding this optimization.