Remove default slicing rules on choice elements
When a choice element adheres to the default slicing rules provided by
SUSHI, including those rules is not necessary as long as one of the
choices exists. Therefore, if there is a rule with a path to one of
those choices, slicing rules on the choice element can be removed if
they have the default value applied.