Extract ObeysRule and Invariant
When an element has a constraint array, extract one obeys rule for the
element and one invariant for each constraint. In FSH, a single obeys
rule may contain multiple constraints, but in SUSHI, an obeys rule
contains only one constraint. Therefore, an ExportableObeysRule
maintains a list of constraint keys.

Maintain invariants along with the profile or extension that refers to
them in order to keep them grouped together when exporting FSH.