Process Invariants in a separate Processor class
Invariants are no longer created by ObeysRuleExtractor. If a
StructureDefinition is successfully processed by ProfileProcessor or
ExtensionProcessor, get the invariants from the StructureDefinition
using InvariantProcessor. ObeysRuleExtractor marks the paths that are
used by InvariantProcessor as processed so that those paths do not
result in the extraction of CaretValueRules.

Only one Invariant with a given name may be present in the Package. If
an Invariant with the same name would be added, do not add it. If the
new Invariant has a different definition than the Invariant with the
same name, log an error.