Remove default 0..0 rules from extensions
When SUSHI exports extensions, it creates default 0..0 rules for extension or value[x], whichever path was not used by the extension.  This is because FHIR says that you can only use extension or value[x], but not both.

Now if GoFSH detects that type of situation, it will remove the 0..0 rules since SUSHI should generate them automatically (and they are therefore redundant).