Generate configuration YAML from ImplementationGuide (#2)
* Base definition for configuration processing

ConfigurationProcessor uses an ImplementationGuide resource to create an
ExportableConfiguration instance. A package may only have one such
instance. The ExportableConfiguration instance is responsible for
generating the YAML to represent a valid SUSHI configuration file.

* Include optional fields when generating config

The id, name, status, and version fields are optional. If they are
present in the ImplementationGuide resource, they will be added to the
configuration. Otherwise, they are skipped. No message is logged when
skipping an optional field.

* Use stricter regex for canonical

* Add tests for adding configuration to package