Extract ObeysRule and Invariant (#4)
* Extract ObeysRule and Invariant

When an element has a constraint array, extract one obeys rule for the
element and one invariant for each constraint. In FSH, a single obeys
rule may contain multiple constraints, but in SUSHI, an obeys rule
contains only one constraint. Therefore, an ExportableObeysRule
maintains a list of constraint keys.

Maintain invariants along with the profile or extension that refers to
them in order to keep them grouped together when exporting FSH.

* Process Invariants in a separate Processor class

Invariants are no longer created by ObeysRuleExtractor. If a
StructureDefinition is successfully processed by ProfileProcessor or
ExtensionProcessor, get the invariants from the StructureDefinition
using InvariantProcessor. ObeysRuleExtractor marks the paths that are
used by InvariantProcessor as processed so that those paths do not
result in the extraction of CaretValueRules.

Only one Invariant with a given name may be present in the Package. If
an Invariant with the same name would be added, do not add it. If the
new Invariant has a different definition than the Invariant with the
same name, log an error.

* Remove InvariantProcessor, add InvariantExtractor

In order to more consistently apply processors and extractors,
Invariants are produced by the InvariantExtractor. Processors should
accept input of a type based on their name and provide multiple types of
output. Extractors should accept various inputs and provide output of a
type based on their name.

* Reuse elements between extractors

A ProcessableElementDefinition keeps track of which of its paths have
been processed. Instantiate these in the processor, and pass them to any
extractors that operate on elements.

* Constraints on different elements may share a key

FSH treats Invariants as global entities, so only one Invariant with a
given key is created. Only create ObeysRules in cases where the
associated Invariant would provide the correct values. Otherwise,
specify constraints using CaretValueRules.

Do not emit an error when more than one constraint on the same element
has the same key. GoFSH is not a FHIR validator.

* Find Invariants with matching key on current element

* Export Invariants in FSHExporter