Improve inference of name and id from canonical
A valid canonical URL can take many forms. The function to use the
canonical URL to determine the name and id for a configuration now
correctly handles a greater variety of URLs. Test cases are added to
demonstrate these possibilities. The code for inferring the name and id
is moved to a separate function for ease of testing.

In ConfigurationExtractor, clean up imports and add an explanation to
the relatively dense getStringMode function. Use default values for
canonical and FHIRVersion when inferred values are unavailable.