Extract configuration before processing resources
Having the configuration available while processing resources allows
processors to check configured values. This allows the processors to
avoid adding unnecessary rules by detecting scenarios where the effect
of the configuration would be the same as the rule.

In order to implement this, the strategy for getResources is changed.
Files are first registered and sorted by resourceType. The raw file
contents are used to extract a configuration. Then, the file contents
are processed to create exportable FSH objects.