Detect and replace FHIR core resource urls (#25)
* Detect and replace FHIR core resource urls

When a FHIR core resource url is a profile parent or the type in an
OnlyRule, it is not necessary to use the full url. The name of the
resource is sufficient.

* Detect core FHIR resources as parents of Extensions

* Preserve core FHIR urls in parents to avoid name collisions

If a Profile or Extension's parent is a core FHIR resource, and that
FHIR resource shares a name with a local StructureDefinition, do not
change the url to a name.

* Replace only rule type urls with local StructureDefinition names

If a type constraint refers to a local StructureDefinition, refer to it
by name instead of the full url. If a type constraint refers to a core
FHIR resource that shares a name with a local StructureDefinition, do
not modify the type constraint.

* Combine if statements