Minor fixes to configuration processing and extraction
Configurations should always set FSHOnly to true.
Log a warning when an input file is valid JSON, but does not have a
recognized resourceType.
FHIRProcessor.process returns a new Package, rather than mutating a
Package provided as a parameter.
Add test for FHIRProcessor correctly using ConfigurationProcessor on a
file with resourceType ImplementationGuide.