Extract configuration from non-IG resources (#20)
* Extract Configuration with canonical

Resources in the package are checked for a url directly on the resource.
If they follow the standard SUSHI pattern, the correct canonical url is
likely to be found.

* Infer additional attributes when extracting configuration

The fhirVersion, version, and status attributes can be directly inferred
from rules on resources. The id and name attributes can be inferred
based on the canonical.

* Improve inference of name and id from canonical

A valid canonical URL can take many forms. The function to use the
canonical URL to determine the name and id for a configuration now
correctly handles a greater variety of URLs. Test cases are added to
demonstrate these possibilities. The code for inferring the name and id
is moved to a separate function for ease of testing.

In ConfigurationExtractor, clean up imports and add an explanation to
the relatively dense getStringMode function. Use default values for
canonical and FHIRVersion when inferred values are unavailable.

* Extract configuration before processing resources

Having the configuration available while processing resources allows
processors to check configured values. This allows the processors to
avoid adding unnecessary rules by detecting scenarios where the effect
of the configuration would be the same as the rule.

In order to implement this, the strategy for getResources is changed.
Files are first registered and sorted by resourceType. The raw file
contents are used to extract a configuration. Then, the file contents
are processed to create exportable FSH objects.

* Minor fixes to configuration processing and extraction

Configurations should always set FSHOnly to true.
Log a warning when an input file is valid JSON, but does not have a
recognized resourceType.
FHIRProcessor.process returns a new Package, rather than mutating a
Package provided as a parameter.
Add test for FHIRProcessor correctly using ConfigurationProcessor on a
file with resourceType ImplementationGuide.

* Include file path in errors from FHIRProcessor.process

* Include FSHOnly in exported configuration

* Don't include hl7 and fhir at the start of name and id

* Include hl7 and fhir in id, but not name