Extract ValueSetConceptComponentRule (#35)
* Extract ValueSetConceptComponentRule

A ValueSet may contain included or excluded components that can be
defined by as a FSH ValueSet concept component. These components must
contain at least one concept definition, and therefore must also have a
system specified. Other components are defined by a FSH ValueSet filter
component, which are outside the scope of this extractor.

* Extract ValueSetConceptComponentRule for invalid concepts

A concept without a code violates FHIR spec, but goFSH is not a FHIR
validator. Create the invalid rule so that the error is brought to the
user's attention.

Compact the list of rules in ValueSetProcessor only after all rules have
been extracted.

* Extract rules when processing ValueSet

* Remove type check for ValueSetComponent