Prefer Quantity-style rules
When code, system, unit, and value are all present on a quantity, it is
possible to combine (code, system, unit) or (code, system, value),
leaving the remaining rule unchanged. Prefer combining (code, system,
unit), because this outputs a rule of the recommended style. Note that
this combination is only possible when the system is the standard UCUM
url. If the system is some other value, combine (code, system, unit).