Move rules on CodeableConcept.coding[0]
SUSHI allows a FSHCode value to be assigned to an element of type
CodeableConcept. When this occurs, the value is put into the first
position of the element's coding array. Therefore, when rules are
combined for the first element of a coding array, change the rule's path
to be directly on the CodeableConcept element as long as there are no
other rules present that would make such a change potentially confusing.