ValueSetComponentRules may export to multiple lines (#46)
* ValueSetComponentRules may export to multiple lines

When the FSH export from an ExportableValueSetComponentRule is more than
100 characters long, break it up into multiple lines. Each line in a
filter component will begin with "and" or "where". Each line in a
concept component will begin with "and" or "from".

* Improve linebreak logic

Adding linebreaks after constructing the FSH string led to problems when
the FSH contained strings that matched the regular expression that added
linebreaks. Instead, rebuild the components of the FSH string using
linebreaks as separators.

* Modify style for simple concept components, indentation

When a concept component is only from a system, write one rule for each
code.
Successive lines of a rule are indented by four spaces. Prefer putting
"and" at the end of a line, rather than the beginning.

* More tests with exclude