Combine CaretValueRules and AssignmentRules for Coding and Quantity types (#38)
* Combine Coding and Quantity CaretValueRules on Profiles and Extensions

A set of CaretValueRules on a Profile or Extension may define the
elements of a value of type Coding or Quantity. Combine these rules to
use the FshCode or FshQuantity types to define the value on the parent
element with a single rule.

* Combine Coding and Quantity AssignmentRules on Instances

These follow similar logic to the CaretValueRules on Profiles and
Extensions, except that the path is checked instead of the caret path,
and the logic pertaining to the separately checking the non-caret-path
is inapplicable.

* Add comment regarding Profile and Extension assignment rules

The optimizer specifically does not run on assignment rules in Profiles
and Extensions. The comment explains why this is unnecessary.

* Prefer Quantity-style rules

When code, system, unit, and value are all present on a quantity, it is
possible to combine (code, system, unit) or (code, system, value),
leaving the remaining rule unchanged. Prefer combining (code, system,
unit), because this outputs a rule of the recommended style. Note that
this combination is only possible when the system is the standard UCUM
url. If the system is some other value, combine (code, system, unit).

* Move rules on CodeableConcept.coding[0]

SUSHI allows a FSHCode value to be assigned to an element of type
CodeableConcept. When this occurs, the value is put into the first
position of the element's coding array. Therefore, when rules are
combined for the first element of a coding array, change the rule's path
to be directly on the CodeableConcept element as long as there are no
other rules present that would make such a change potentially confusing.

* Style consistency in helper functions

Check for self, then type, then other conditions when finding other
siblings.
Checking with some is more direct than checking with
filter.length.