Process unsupported CodeSystems as Instances (#48)
* Process unsupported CodeSystems as Instances

If a CodeSystem uses paths that cannot easily be represented using FSH CS syntax, process it as an Instance instead.

* Don't Resolve URLs of Unsupported CS/VS

Since SUSHI doesn't properly resolve CS/VS names when they are FSH instances, don't resolve the URLs to names in GoFSH.  This can be revisited later when SUSHI is fixed.

This required a refactor to MasterFisher to allow fishing specifically only in the lake.

* Remove commented out code

* Fix merge issues

Last merge brought in some changes that required updates in this branch.

* Don't over-optimize things that look like codes but aren't!

The CombineCodingAndQuantityValuesOptimizer was too aggressive in optimizing; optimizing some things that looked like codes but weren't actual FHIR code/Coding/CodeableConcept types (for example, CodeSystem.concept).  Now the optimizer checks the actual type before optimizing.

* Only lookup elements on SDs w/ snapshots

When determining types for an element path, if the SD does not have a snapshot, try to use the SD of it's base type instead.

This should be fine since profiles can't really introduce brand new paths (they can only constrain).