Avoid redefining implicit slice elements
When slices are defined on a nested element, any slices on an element
closer to the root will automatically contain the nested slices. These
elements can also be inherited, and further slices near the root may be
added. When processing elements that are slices, it is only necessary to
create a ContainsRule for that element if the slice will not be
automatically created.

Since the above is basically spaghetti, here's an example:
Parent profile states that element at path A.B contains a slice
"banana". That element's id is A.B:banana. This would be represented
with the contains rule "A.B contains banana".
Child profile states that element at path A contains a slice "apple".
That element's id is A:apple.
The child profile needs a contains rule that says "A contains apple", as
this is a completely new slice.
The child profile does not need a contains rule that says "A[apple].B
contains banana", because this can already be inferred based on the
slice defined on the parent profile.