AssignmentRule quantity must have UCUM system
When extracting assignment rules for a quantity or ratio (which contains
quantities), construct a FSHQuantity instance only if the system is the
url for UCUM. This is because SUSHI currently assumes that quantities
and ratios use this system. If the system is not the UCUM url, create
separate rules for the code and the value.