Handle assignment to quantities without values
A fixed or pattern Quantity on an element may not have the value field
present. If the code field is present, it is possible to create an
assignment rule that assigns a FshCode to the Quantity. If code is not
present, no assignment rules can be created, and caret value rules will
be used.