Extract assignment rules for parts of Ratios
The pattern or fixed value on an element of type Ratio may not always be
possible to represent as a FshRatio. When this occurs, extract rules for
the numerator and denominator separately. These elements are of type
Quantity, so apply the existing logic for how to extract assignment
rules on Quantity elements.