AssignmentRule quantity must have UCUM system (#70)
* AssignmentRule quantity must have UCUM system

When extracting assignment rules for a quantity or ratio (which contains
quantities), construct a FSHQuantity instance only if the system is the
url for UCUM. This is because SUSHI currently assumes that quantities
and ratios use this system. If the system is not the UCUM url, create
separate rules for the code and the value.

* Handle assignment to quantities without values

A fixed or pattern Quantity on an element may not have the value field
present. If the code field is present, it is possible to create an
assignment rule that assigns a FshCode to the Quantity. If code is not
present, no assignment rules can be created, and caret value rules will
be used.

* Extract assignment rules for parts of Ratios

The pattern or fixed value on an element of type Ratio may not always be
possible to represent as a FshRatio. When this occurs, extract rules for
the numerator and denominator separately. These elements are of type
Quantity, so apply the existing logic for how to extract assignment
rules on Quantity elements.