Fix processing test when current working directory has gofsh folder
The test was creating a gofsh folder in the current working directory -- so if it already existed, the user prompt was activating but no one was answering it.

Really, we should be changing the working dir to the temporary folder so the test is not mucking with a real gofsh folder in the project directory.