Fix mapping indices
Since mapping arrays in the differential can be subsets of the corresponding arrays in the snapshot, GoFSH's first attempt at assigning indices is relative to the differential.  When we process affected caret rules, we should attempt to fix the index so it matches the correct mapping in the snapshot element, since that is what SUSHI will use when applying the caret rule.

If there is no snapshot available, we cannot correct the index, so we should emit a warning and also output a comment in the FSH indicating that the mapping index may be incorrect.