Fix constraint and mapping indices (#89)
* Fix processing test when current working directory has gofsh folder

The test was creating a gofsh folder in the current working directory -- so if it already existed, the user prompt was activating but no one was answering it.

Really, we should be changing the working dir to the temporary folder so the test is not mucking with a real gofsh folder in the project directory.

* Fix constraint indices

Since constraint arrays in the differential can be subsets of the corresponding arrays in the snapshot, GoFSH's first attempt at assigning indices is relative to the differential.  When we process affected caret rules, we should attempt to fix the index so it matches the correct constraint in the snapshot element, since that is what SUSHI will use when applying the caret rule.

If there is no snapshot available, we cannot correct the index, so we should emit a warning and also output a comment in the FSH indicating that the constraint index may be incorrect.

* Do not output caret rules for constraint.source (unless necessary)

SUSHI autopopulates constraint.source to the url of the current profile, so unless the constraint.source is some other URL, we don't need to output a caret rule for it.

* Fix mapping indices

Since mapping arrays in the differential can be subsets of the corresponding arrays in the snapshot, GoFSH's first attempt at assigning indices is relative to the differential.  When we process affected caret rules, we should attempt to fix the index so it matches the correct mapping in the snapshot element, since that is what SUSHI will use when applying the caret rule.

If there is no snapshot available, we cannot correct the index, so we should emit a warning and also output a comment in the FSH indicating that the mapping index may be incorrect.