Optimize Quantity values based on new FSH STU2 syntax
New FSH syntax for Quantity values in assignment and caret value rules
allow for more elements to be combined into a single rule. The system on
the Quantity is no longer required to be UCUM. The Quantity can now
contain the display string in its unit element. These changes allow the
ExportableAssignmentRule and ExportableCaretValueRule to take advantage
of the new implementation of FshQuantity.toString without any further
changes. As such, no tests are added for those two classes.