Optimize Quantity values based on new FSH STU2 syntax (#105)
* Optimize Quantity values based on new FSH STU2 syntax

New FSH syntax for Quantity values in assignment and caret value rules
allow for more elements to be combined into a single rule. The system on
the Quantity is no longer required to be UCUM. The Quantity can now
contain the display string in its unit element. These changes allow the
ExportableAssignmentRule and ExportableCaretValueRule to take advantage
of the new implementation of FshQuantity.toString without any further
changes. As such, no tests are added for those two classes.

* Support new Quantity syntax in AssignmentRuleExtractor

The AssignmentRuleExtractor does not need to check a quantity's system
in order to create a FSHQuantity to use in an assignment rule. This
applies to quantity and ratio values.

* Remove UCUM vs non-UCUM tests

Quantities are processed the same way regardless of their system.
Remove the comment indicating that switchQuantityRules should be removed
from InstanceProcessor. This function should still be used in order to
guarantee that correct FSH is produced without optimizers.