Extract rules for Logical and Resource entities
A StructureDefinition can represent one of four FSH types. If
StructureDefinitionProcessor is extracting rules for a Logical or
Resource, the AddElementRuleExtractor will be used. The root element
does not need to be added, but is still processed so that the element's
paths are marked as processed.

Logical and Resource have default parents, so if no parent is specified
by the input, use the defaults.