Update SUSHI dependency to 2.0.0 beta (#128)
* Update SUSHI dependency to 2.0.0 beta

The new major version of SUSHI contains a lot of new features relevant
to goFSH. FSH types and rules now have toFSH methods, which makes some
code in goFSH redundant. ExportableValueSetComponentRule needed to be
changed immediately due to the use of a private function, and other
classes will be changed in later commits.

Modify the mock of loadDependency used in Processing tests. The previous
mock was failing due to the fhirdefs object not being able to set a
value on its loadDependency property.

* Move mock of loadDependency so that it actually gets used

Mocks of included modules require somewhat more advanced rituals in
order to actually get used. This mock rebuilds the fsh-sushi module with
only the loadDependency function changed.