LakeOfFHIR identifies Logicals and Resources
Add logic in LakeOfFHIR in order to identify StructureDefinitions that
represent these FSH types. The StructureDefinitionProcessor is given
StructureDefinitions that represent Profiles, Extensions, Logicals, and
Resources. The InstanceProcessor receives all other
StructureDefinitions.

Update StructureDefinitionProcessor to more accurately determine which
FSH type to create, and to throw an error when receiving a
StructureDefinition that does not represent one of those FSH types.