Upgrade SUSHI dependency to 2.0.0, refactor toFSH implementations (#130)
* Update SUSHI dependency to 2.0.0 beta

The new major version of SUSHI contains a lot of new features relevant
to goFSH. FSH types and rules now have toFSH methods, which makes some
code in goFSH redundant. ExportableValueSetComponentRule needed to be
changed immediately due to the use of a private function, and other
classes will be changed in later commits.

Modify the mock of loadDependency used in Processing tests. The previous
mock was failing due to the fhirdefs object not being able to set a
value on its loadDependency property.

* Remove toFSH implementation from ExportableAssignmentRule

SUSHI now provides this function's implementation. The SUSHI
implementation fshifies string values, so do not fshify string values
when extracting an ExportableAssignmentRule.

* Remove toFSH implementation from ExportableBindingRule

* Remove toFSH and cardToString implementations from ExportableCardRule

* Refactor goFSH implementation in ExportableCaretValueRule

The goFSH method provided by SUSHI includes everything except
fshComment, since that field is added by goFSH.

* Remove toFSH implementation from ExportableCodeSystem

* Remove toFSH implementation from ExportableConceptRule

* Remove toFSH and flagsToString implementations from ExportableFlagRule

* Remove toFSH implementation from ExportableInstance

* Remove toFSH implementation from ExportableInvariant

* Refactor toFSH method in ExportableMapping

* Remove toFSH implementation from ExportableMappingRule

* Remove additional toFSH implementations

toFSH implementations are removed from ExportableOnlyRule,
ExportableProfile, ExportableValueSet, and
ExportableValueSetComponentRule types. Additional functions that
supported these implementations are also removed.

* Refactor toFSH in ExportableExtension

The Extension class provided by fsh-sushi has a toFSH method that is
almost correct. It can be made correct by removing the parent keyword
when the default parent is used. Since this function was the last place
that used metadataToFSH, metadataToFSH is removed.

* Do not fshify string when getting FSH value

Values that are obtained from getFSHValue are used in CaretValueRules
and AssignmentRules. These rules will fshify string values as part of
their toFSH method. Remove the fshifyString call in order to avoid
double-fshification.

* Remove commented-out code