Warn on slices and assigned values on Logical and Resource elements
FHIR does not allow slices to be defined or values to be assigned on
newly defined elements on Logicals and Resources. The rules for these
are still created and exported, but the user is warned that FHIR will
not allow it.

Various fixes to places with type mistakes where a Logical was expected,
but a Resource was created instead.