Fix logic for determining if SD represents Instance
The previous logic in LakeOfFHIR was too strict about what kinds of
StructureDefinitions could represent a Profile. Any kind other than
"logical" is allowed on a Profile. The function that implements this
logic is renamed to isSDForInstance to try to more clearly indicate that
this function should return true when given an SD that represents an
Instance (and not a Profile, Extension, Logical, or Resource), and
should return false in all other cases.

The logic in StructureDefinitionProcessor is also updated to allow
Profiles on complex-type or simple-type.

A Profile on a primitive-type is added to the LakeOfFHIR tests in order
to demonstrate that such a StructureDefinition will be identified as a
Profile. A Profile on a complex-type is added to the
StructureDefinitionProcessor tests in order to demonstrate that having a
type of "Extension" is necessary for an Extension, and that other
StructureDefinitions with kind "complex-type" and derivation
"constraint" will result in a Profile.