Do not add caret or assignment rules with empty values
When input JSON contains an empty list, an empty object, or an explicit
null or undefined, do not create a rule that uses that empty value.
These values are not valid FHIR, and would result in invalid FSH.
Display an error message to the user so they know the value is being
ignored.