Remove sea creature utility methods
With this commit, we bid a fond farewell to our longtime supporters and
accomplices, the sea creatures listed with the results when running
goFSH. Through resources compliant and noncompliant, the sea creatures
were there for us, ready to appear and report their current moods. For
this service to us, we remain gratitudinous, now and always.

For those concerned with the welfare of these sea creatures no longer in
employ, allow me to assure you, that you may put your hearts at ease: no
sea creatures have been harmed, and all have found gainful employment at
positions matching their skills and experience gained. We hope that
their time working as part of the command line interface will be fondly
remembered, and of benefit to them, as they continue to sojourn as we
all do, through the rolling waves and fathomless depths of life.

And should it arise that we once again require the services of sea
creatures, they will return! Oh, with such fanfare and triumph we should
welcome them back, should we need them once more, to make consistent and
delightful the command line interface output formatting. They remain in
good health, and are living their best lives, as best they are able. But
for now, we say to them good-bye, and wipe the tears from our eyes as we
do.