Add Rule Path Contexts optimizer
This new optimizer attempts to make use of context paths in rules. If a
rule's path can be partially or entirely represented by preceding
rule's path context, set an indent on the current rule and shorten its
path to only contain the parts that come after the established context.

This new optimizer needs to run after all other optimizers, including
the array indexing optimizer. Several optimizers have their ordering
specifications changed to enforce this.

The toFSH methods on rules are not yet updated.