Track enabled optimizers when loading
When loading optimizers, some optional optimizers may not be enabled.
Other optimizers may reference them in their runBefore or runAfter
properties. If an optimizer that will run is configured to run before
the disabled optimizer, this does not represent an error case. However,
if an optimizer that will run is configured to run after a disabled
optimizer, this represents an error case. Show error messages specific
to whether an optimizer listed as an adjacency isn't found or is turned
off.