Generate Logicals and Resources (#132)
* Create Logical and Resource from StructureDefinition

These two new FSH types are created from FHIR StructureDefinitions.
Update test fixtures that are supposed to create Profiles so that they
meet the new requirements for profiles.

* Extract rules for Logical and Resource entities

A StructureDefinition can represent one of four FSH types. If
StructureDefinitionProcessor is extracting rules for a Logical or
Resource, the AddElementRuleExtractor will be used. The root element
does not need to be added, but is still processed so that the element's
paths are marked as processed.

Logical and Resource have default parents, so if no parent is specified
by the input, use the defaults.

* LakeOfFHIR identifies Logicals and Resources

Add logic in LakeOfFHIR in order to identify StructureDefinitions that
represent these FSH types. The StructureDefinitionProcessor is given
StructureDefinitions that represent Profiles, Extensions, Logicals, and
Resources. The InstanceProcessor receives all other
StructureDefinitions.

Update StructureDefinitionProcessor to more accurately determine which
FSH type to create, and to throw an error when receiving a
StructureDefinition that does not represent one of those FSH types.

* Add tests for AddElementRuleExtractor

* Warn on slices and assigned values on Logical and Resource elements

FHIR does not allow slices to be defined or values to be assigned on
newly defined elements on Logicals and Resources. The rules for these
are still created and exported, but the user is warned that FHIR will
not allow it.

Various fixes to places with type mistakes where a Logical was expected,
but a Resource was created instead.

* Add Logicals and Resources to FSHExporter

These entity types are included in the various export methods provided
by FSHExporter.

Because Logical and Resource both have a default parent, omit it from
the FSH produced when the parent has the default value. This works
similarly to the logic for Extension.

* Remove generated text from Logicals and Resources

RemoveGeneratedTextRulesOptimizer will check Logical and Resource
objects for removable text rules.

Simplify the type definition for ConstrainableEntity.rules.

* Fix logic for determining if SD represents Instance

The previous logic in LakeOfFHIR was too strict about what kinds of
StructureDefinitions could represent a Profile. Any kind other than
"logical" is allowed on a Profile. The function that implements this
logic is renamed to isSDForInstance to try to more clearly indicate that
this function should return true when given an SD that represents an
Instance (and not a Profile, Extension, Logical, or Resource), and
should return false in all other cases.

The logic in StructureDefinitionProcessor is also updated to allow
Profiles on complex-type or simple-type.

A Profile on a primitive-type is added to the LakeOfFHIR tests in order
to demonstrate that such a StructureDefinition will be identified as a
Profile. A Profile on a complex-type is added to the
StructureDefinitionProcessor tests in order to demonstrate that having a
type of "Extension" is necessary for an Extension, and that other
StructureDefinitions with kind "complex-type" and derivation
"constraint" will result in a Profile.

* Add Logicals and Resources to result printout

* Remove sea creature utility methods

With this commit, we bid a fond farewell to our longtime supporters and
accomplices, the sea creatures listed with the results when running
goFSH. Through resources compliant and noncompliant, the sea creatures
were there for us, ready to appear and report their current moods. For
this service to us, we remain gratitudinous, now and always.

For those concerned with the welfare of these sea creatures no longer in
employ, allow me to assure you, that you may put your hearts at ease: no
sea creatures have been harmed, and all have found gainful employment at
positions matching their skills and experience gained. We hope that
their time working as part of the command line interface will be fondly
remembered, and of benefit to them, as they continue to sojourn as we
all do, through the rolling waves and fathomless depths of life.

And should it arise that we once again require the services of sea
creatures, they will return! Oh, with such fanfare and triumph we should
welcome them back, should we need them once more, to make consistent and
delightful the command line interface output formatting. They remain in
good health, and are living their best lives, as best they are able. But
for now, we say to them good-bye, and wipe the tears from our eyes as we
do.

* Remove default value caret rules from Logical and Resource

SUSHI provides useful defaults for Logical and Resource that do not need
to be provided by caret value rules. Define optimizers to remove those
rules.

* Remove publisher derived dates from Logicals and Resources

* Resolve parent URLs for Logicals and Resources

* Simplify array indexing for Logicals and Resources

Caret rules on Logicals and Resources may be able to take advantage of
soft indexing.

* Instances can be grouped with Logicals and Resources in output

When the output style is group-by-profile, Instances can be grouped with
Logicals and Resources the same way that they can be grouped with
Profiles. The rules for deciding how the Instances are grouped are the
same as before. We just now make groups for Logicals and Resources.

* Clarify comments for group by profile strategy

When using the group by profile strategy for export, each profile,
logical, and resource will get its own file. Examples are only
applicable to profiles and resources, but there isn't a need to change
the logic to support that fact.

* Update installed dependencies

* Use BackboneElement as default type for AddElementRule

If AddElementRuleExtractor finds no types for an element, give it a type
of BackboneElement, as this is often correct. This represents invalid
FHIR input, so show a warning indicating that this default has been
chosen.

* Extract rules for some values on a Logical or Resource root element

The root element on a Logical Resource does not need an AddElementRule.
However, it may have values set on it that need to be represented with a
CaretValueRule. Ignore the cardinality values on this element, but
extract rules for other values.

* Change logic for detecting root element in Logical and Resource

The path of the root element will never contain a . character, so use
that to check for it.

* Use indexOf instead of includes, for compatibility