Set default indent for ExportableRule implementations
Rule classes that implement ExportableRule and use the indent property
now assign the value of a default of 0. This helps avoid having to check
for undefined.

Change name of OptimizerPlugin's toggle function to "isEnabled", as this
more strongly hints that it will return a boolean that indicates whether
the function is enabled or not.