Move optimizer enable check out of loadOptimizers to getResources
All optimizers present in the file system are returned from
loadOptimizers, regardless of whether or not they are enabled. The check
for whether an optimizer is enabled is moved to immediately before the
optimizer will be run. Errors are not logged when a preceding optimizer
is not enabled. Essentially, the principle to keep in mind is that no
optimizer is ever required to be run. As long as the correct ordering is
followed, output should remain valid after running all enabled
optimizers. Tests regarding enabling optimizers are moved to the
Processing module test suite.

Rename optimizer options to processing options, as they may be used for
other parts of processing.